Add keyboard commands to watch mode (#7)
This commit makes watch mode interactive. It now understands these keys:

- Press ENTER to run all tests
- Press "q" to quit

To do this, I start another background thread that listens for key
presses. When a key is pressed, the thread posts a `:keypress` event to
the event loop. The event loop then takes an appropriate action.

Because they keypress listener is running in a thread separate from the
file system listener, that means the watcher is still able to detect
changes to files and auto-run tests. The event loop funnels all of these
events into a single thread of execution.