Add interactive watch command for running tests based on git status
The `git ls-files` command allows us to get a list of all files that
have been added or changed since the last commit. This is a convenient
way to determine what tests to run.

This commit adds a "d" keypress command to the watch mode. When the "d"
key is pressed, we check git for new and changed files, then find the
corresponding test files, and run those.