Ensure watch (d)iff command finds staged files
Previously, when "d" was pressed in watch mode, we were using
`git ls-files` to find the list of new and modified files. If a file was
modified but staged/cached (i.e. it was added with `git add` but not yet
committed), it would not be detected.

This commit changes the diff implementation to use `git status` instead,
which lists both staged and un-staged changes.