Handle minitest flags that accept values, like `--seed` (#19)
mighty_test needs an option parser to handle its own flags, like
`--all`, `--shard`, and `--watch`. However, users can also pass in
Minitest flags like `--seed` or `--fail-fast` and expect them to be
passed to Minitest when tests are run.

This is challenging, because the full range of flags that Minitest
supports isn't known in advanced. Minitest plugins can add arbitrary
flags as runtime. Therefore we can't write a single option parser that
works for mighty_test and Minitest.

The implementation in this commit takes the following approach: First,
parse the flags that mighty_test understands, like `--all`. Then check
if there are any unrecognized flags remaining. If so, use Minitest's
internal `process_args` method to parse the remaining flags.

The result isn't perfect (Minitest's error handling does a hard exit
when it encounters at invalid option, so we don't have an opportunity to
format or augment the help message), but it works for common cases.

I removed `optparse` because its design assumes a single parser that
handles the entirety of ARGV. We need a parser that can decode flags one
at a time, and then pass the remainder to Minitest, something that I
couldn't get to work with `optparse`.

I decided to "brute force" the parsing of the few flags that mighty_test
understands, rather than pull in a third-party gem.