Allow use of interactive debugger when running tests in watch mode (#30)
Before, `mt --watch` would take full control of stdin to listen for keyboard
commands. This prevented the use of an interactive debugger when a test is 
being executed.

This PR changes the event loop implementation so that control of stdin is 
relinquished when running tests. This allows an interactive debugger to be
used while tests are running.

Specifically:

- Move stdin reader (i.e. `$stdin.getc`) out of a background thread and into
  the main event loop.
- `$stdin.getc` blocks until a character is typed.
- When the file system listener background thread detects a change, raise an 
  exception in the main event loop to break out of `$stdin.getc` and proceed
  to run the tests.